Revamped the crafting handler for upgradable items.
The recipes aren't in yet, but this is the groundwork for handling
VoidUpgradable items. I decided against the interface and opted for a
simpler crafting handler with some tweaked methods to include things
aside from the void satchel. The initial capacity nerf, and subsequent
upgrade recipes are soon to follow. However, for this, I will be
incorporating NEI's API, so that I can hide the upgrade recipes this
adds to the recipe registry.