Small commit!
Removed the derpy console print in AlkahestHandler that once upon a time
I added while debugging. Apparently it spams the eff out of server logs.

Er.. Indentation in a new item? Ignore this.

ItemXR had a derpy reversed sign, and also needed to cast an int as a
short. (This should fix the cooldown bug)

Names updated to say "Elixir" instead of "Elixer", this was an old
spelling error that I almost forgot about.