Removed Alkahest (for the time being at least)
Is Redstone really that much of a rarity that we need that?
With many modpacks where lots of Redstone is needed (large tech
modpacks), there is already many options for renewability. I took a
long time thinking about this, and it also seems in many mods
Cobble/Sand/Dirt are interchangable. So it would probably be possible
to get an infinite amount of Redstone, hence infinite dupes. (untested)

I also positioned the gun sprite, since it dosen’t really get rendered
in-hand anymore.