Started work on the fundementals of Alkahestry
- The Tomb of Alkahest now has a buffer of 256 (configurable) Redstone
inside itself. The recipies don’t work, but you can ‘charge’ up the
Tomb by placing Redstone or a Redstone Block alongside it in a crafting
table.

This is a new feature, and that means new bugs. So, if you could help
me stomp the bugs via the Github issue tracker, that would be great!