In the name of abstraction...
Almost complete. Also, we are back to error status.

- I thought it was a little redundant to have 50 calls for item object
init with no parameters for any of them. So I talked to x3n0, and from
now on we are abstracting away item/block registration with a XRInit
annotation. The AbstractionHandler class (bound to be renamed shortly)
scans the ‘xreliquary.blocks’ and ‘xreliquary.item’ packages for
classes, checks if they have XRInit, then registers them automagically.
It’s quite nifty.