Medallion Fix and HUD Configs
Added config option to reposition the handgun HUD (ripped from EE3 1.6.4
git), its options are the four corners to get us going. 0 - top left, 1
- top right, 2 - bottom left, 3 - bottom right (default), I haven't
tested the layouts in all the positions, will need to do that.

More OCD on the HUD, not much else.

Fixed some derping in the hero medallion, it appears to drain properly
now (it was absorbing infinitely and not reducing levels). Still need to
test that it gives exp back correctly/levels the player without derps.