Huge Commit.
1) Blaze shot now deals fire damage and sets the mob on fire before
doing damage so that meat will cook when dropped.
2) Ender staff projectile can no longer send you to the void by mistake.
3) Sojourner's Staff now has a HUD (keep reading).
4) Removed Wraith's Eye (no longer exists). Now the Ender staff does
what the Wraith's Eye used to.
5) Configs for Ender Pearl cost and Node Warp cost, and Ender pearl
worth.
6) Sojourner's Staff can now cycle through the item types it contains
(shift left click) and no longer has a damage bar. It has an internal
storage for each item type in the list, defaults to 1500 but
configurable.
7) Made a bunch of stuff ItemToggleable. Basically all the things.
8) Midas Touchstone now has a config list of additional items that can
be added for repair. It now has configurable cost/glowstone dust worth.
9) Lang file adjustments.
10) Ender staff sprite got some love, it's now animated.
11) Lantern of Paranoia now has a recipe. It also got some fixes because
it was stupid about placing torches in water, and other things. Also got
adjustments to work with the Sojourner's staff in its new style.