Derp
Removed dirt. I'm dumb.
Forgot to mention last commit, changed toml to cfg for the config file
so that common config scrapers could parse our config easier without
being modified.
Also made it so that "minecraft:torch" isn't in the config for the
Sojourner's, and cannot be removed (so that the Lantern can't be borked
by people derping with configs)