New Items and Sprites
Refactor of existing entities for easier-to-find layout.
Updated a bunch of sprites to be shinier/have animations.
Added gust.ogg for eventual addition of a sound effect.
Tweaked the lang file.
Finalized Rod of Lyssa and Pyromancer staff, which are hilarious and
awesome.
More progress on the Infernal Chalice, made design decision to remove an
OP ability.
Made a few methods less obfuscatey.