Mechanics/Rebalance of potions
Potion map heavily rebalanced, hopefully for the better. Mechanics of
potion mixing have changed, you no longer have to have amp 1 or higher
to create a potion. Any combination with matching effects will produce
at least a 0 amp potion. Potion effects that don't exist in at least 2
or more of the ingredients will no longer appear in the result potion.
Duration formula now more generous with the result of two mixing
ingredients (/1.2F instead of 1.5). Most mappings are now lower level to
compensate for the fact that now instead of choosing the higher of two
or more ingredients, the effect is additive. No ingredients in the
"vanilla" mapping are higher than a +1 amp, meaning a level II potion is
as high as you can go without mixing essences (intended), to a maximum
amplifier of 4 (level V potion). I may adjust or develop a formula table
for combining essences, to make it more difficult to stack
stupid-powerful ones, but I'm going to leave it as is for a while.