Side-Only Types of Fixes and Minor Improvements
Prevents server crashes from a part of the bullet homing effect, as well
as cauldron particle effect crashes.  The cauldron now only takes what
it can use from item stacks thrown into it (also meaning that if
multiple redstone are thrown in, it recognizes more than the addition of
just one).  I also commented-out the line adding the unused tooltips for
the empty and standard magazines.